device-state: do not reboot while HUP is in progress
Change-type: patch

+++ src/device-state/index.ts
@@ -681,7 +681,12 @@ export const applyTarget = async ({
 			// For application manager, the reboot breadcrumb should
 			// be set after all downloads are ready and target containers
 			// have been installed
-			if (steps.every(({ action }) => action === 'noop') && rebootRequired) {
+			if (
+				steps.every(({ action }) => action === 'noop') &&
+				rebootRequired &&
+				!(await updateLock.isHUPInProgress()) &&
+				!(await applicationManager.isOsUpdatePending(rawTargetApps))
+			) {
 				steps.push({
 					action: 'reboot',
 				});
